requirements.txt
click==7.1.2
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
importlib-metadata==1.6.0
jsonpickle==1.4.1
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
zipp==3.1.0
